main/api/requirements.txt
aloe==0.2.0
aloe-django==0.2.0
ansicolors==1.1.8
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
cssselect==1.1.0
Django==3.1.4
django-filter==2.4.0
django-nose==1.4.7
djangorestframework==3.12.2
docopt==0.6.2
gherkin-official==4.1.3
idna==2.10
Markdown==3.3.3
nose==1.3.7
prettytable==2.1.0
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
sqlparse==0.4.1
tabulate==0.8.9
urllib3==1.26.3
uWSGI==2.0.19.1
wcwidth==0.2.5
yarg==0.1.9
